refactor(fuzz): split harness into modules, source IDs from spl crates
Moves setup helpers to helpers.rs and tunables to constants.rs. Token
and ATA program IDs plus ATA derivation now come from
spl-token-interface and spl-associated-token-account-interface instead
of hardcoded addresses; the system program ID comes from
solana-sdk-ids.
