+++ fuzz/subscriptions/Cargo.toml
@@ -17,7 +17,11 @@ libafl_bolts = { version = "0.15.1", features = ["std"] }
 solana-clock = "3.0"
 solana-keypair = "3.0"
 solana-pubkey = "3.0"
+solana-rent = "3.0"
+solana-sdk-ids = "3.0"
 solana-signer = "3.0"
+spl-associated-token-account-interface = "2.0.0"
+spl-token-interface = "3.0.0"
 
 [[bin]]
 name = "invariant_test"
